# src/app/watchlist/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "backend for watchlist micro-service",
        "title": "backend",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/api/watch/watchlist": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Handles get, add, delete request for watchlists",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Add, Del, Get of watchlist",
                "parameters": [
                    {
                        "description": "ADD, GET, DEL request of watchlist",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.BFFAdgToWatchlistRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "action completed successfully",
                        "schema": {
                            "$ref": "#/definitions/models.BFFAdgToWatchlistResponse"
                        }
                    },
                    "201": {
                        "description": "Data added successfully",
                        "schema": {
                            "$ref": "#/definitions/models.BFFAdgToWatchlistResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input payload",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorAPIResponse"
                        }
                    },
                    "404": {
                        "description": "User does not exist",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorAPIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorAPIResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.ActionType": {
            "type": "string",
            "enum": [
                "ADD",
                "DEL",
                "GET"
            ],
            "x-enum-varnames": [
                "ADD",
                "DEL",
                "GET"
            ]
        },
        "models.BFFAdgToWatchlistRequest": {
            "type": "object",
            "required": [
                "action",
                "scrip_id"
            ],
            "properties": {
                "action": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.ActionType"
                        }
                    ],
                    "example": "GET"
                },
                "scrip_id": {
                    "type": "string",
                    "example": "JIO_123"
                },
                "watchlist_id": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    },
                    "example": [
                        0
                    ]
                }
            }
        },
        "models.BFFAdgToWatchlistResponse": {
            "type": "object",
            "properties": {
                "action": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.ActionType"
                        }
                    ],
                    "example": "GET"
                },
                "status": {
                    "type": "string",
                    "example": "fetched successfully!"
                },
                "warnings": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "[warnings!]"
                    ]
                },
                "watchlist_with_id": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.WatchlistWithID"
                    }
                }
            }
        },
        "models.ErrorAPIResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "errors": {
                    "$ref": "#/definitions/models.ErrorMessage"
                }
            }
        },
        "models.ErrorMessage": {
            "type": "object",
            "properties": {
                "errorMessage": {
                    "type": "string"
                },
                "key": {
                    "type": "string"
                }
            }
        },
        "models.WatchlistWithID": {
            "type": "object",
            "properties": {
                "watchlist_id": {
                    "type": "integer"
                },
                "watchlist_name": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header",
            "x-extension-openapi": "{\"example\": \"value on a json format\"}"
        }
    }
}